+++ src/VirtoCommerce.Loyalty.Core/Models/LoyaltyOperationLogObject.cs
@@ -1,0 +1,10 @@
+namespace VirtoCommerce.Loyalty.Core.Models;
+
+public class LoyaltyOperationLogObject
+{
+    public string Type { get; set; }
+
+    public string OrderId { get; set; }
+
+    public string OrderNumber { get; set; }
+}
